// questions/src/graphql-codegen-typings.d.ts
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  DateTime: any;
};

export type AdditionalEntityFields = {
  path?: InputMaybe<Scalars['String']>;
  type?: InputMaybe<Scalars['String']>;
};

export type AddQuestionInput = {
  body: Scalars['String'];
  recipient: Scalars['String'];
  title: Scalars['String'];
};
// ...
export type Question = {
  __typename?: 'Question';
  body: Scalars['String'];
  createdAt: Scalars['DateTime'];
  createdBy: Scalars['String'];
  id: Scalars['ID'];
  recipient: Scalars['String'];
  title: Scalars['String'];
};
// ...
export type QuestionDbObject = {
  body: string,
  createdAt: any,
  createdBy: string,
  _id: ObjectId,
  recipient: string,
  title: string,
};